Split series details screen into components
